Convert project into a Maven project
I wanted to add a github-api library to get the number of stargazers on
a git project, and downloading all dependencies manually would be really
annoying, so I just converted the project to a maven project and added
it as a maven dependecy.

Maybe later we can add the other dependencies on the pom file as well.